Renamed protocols (dropping the deprecated keyword). Split GraphQLHttpWebSocket into two classes (one per protocol) that inherit from BaseGraphQLHttpWebSocket
